Refactor handling of global auth0_enabled config
