+++ lib/graphql.ex
@@ -13,6 +13,8 @@ defmodule BridgeEx.Graphql do
   # credo:disable-for-next-line Credo.Check.Refactor.CyclomaticComplexity
   defmacro __using__(opts) when is_list(opts) do
     quote do
+      require Logger
+
       alias BridgeEx.Auth0AuthorizationProvider
       alias BridgeEx.Graphql.Client
 
@@ -28,10 +30,8 @@ defmodule BridgeEx.Graphql do
       @endpoint Keyword.fetch!(unquote(opts), :endpoint)
 
       # optional opts with defaults
+      @auth0_enabled get_in(unquote(opts), [:auth0, :enabled]) || false
       @audience get_in(unquote(opts), [:auth0, :audience])
-      @auth0_enabled if (result = get_in(unquote(opts), [:auth0, :enabled])) == nil,
-                       do: @global_auth0_enabled,
-                       else: result
       @http_options Keyword.get(unquote(opts), :http_options, timeout: 1_000, recv_timeout: 16_000)
       @http_headers Keyword.get(unquote(opts), :http_headers, %{
                       "Content-type" => "application/json"
@@ -92,14 +92,28 @@ defmodule BridgeEx.Graphql do
 
       if @audience && @auth0_enabled do
         defp with_authorization_headers(headers) do
-          with {:ok, authorization_headers} <-
-                 Auth0AuthorizationProvider.authorization_headers(@audience) do
+          with {:ok, authorization_headers} <- get_authorization_headers_for(@audience) do
             {:ok, Enum.into(authorization_headers, headers)}
           end
         end
       else
         defp with_authorization_headers(headers), do: {:ok, headers}
       end
+
+      defp get_authorization_headers_for(audience) do
+        if auth0_enabled_for_app(),
+          do: Auth0AuthorizationProvider.authorization_headers(@audience),
+          else: report_auth0_disabled_error()
+      end
+
+      defp report_auth0_disabled_error do
+        Logger.error("Auth0 is not enabled for this application!
+        To enable it set `config :bridge_ex, auth0_enabled: true` in your config.")
+
+        {:error, "Auth0 not enabled in application"}
+      end
+
+      defp auth0_enabled_for_app, do: Application.get_env(:bridge_ex, :auth0_enabled, false)
     end
   end
 end
